+++ core/modules/modChiffrage.class.php
@@ -72,7 +72,7 @@ class modChiffrage extends DolibarrModules
 		$this->editor_url = 'https://www.atm-consulting.fr';
 
 		// Possible values for version are: 'development', 'experimental', 'dolibarr', 'dolibarr_deprecated' or a version string like 'x.y.z'
-		$this->version = '1.2.1';
+		$this->version = '1.3.0';
 		// Url to the file with your last numberversion of this module
 		//$this->url_last_version = 'http://www.example.com/versionmodule.txt';
 
@@ -563,6 +563,7 @@ class modChiffrage extends DolibarrModules
 		$param = array ( 'options' => array ( 'Chiffrage:chiffrage/class/chiffrage.class.php' => NULL, ), );
 
 		$extrafields->addExtraField('fk_chiffrage', $langs->trans("CHIAssociatedEstimated"), 'link', 100, '', 'propaldet', 0, 0, '', $param, 1);
+		$extrafields->addExtraField('fk_chiffrage', $langs->trans("CHIAssociatedEstimated"), 'link', 100, '', 'projet_task', 0, 0, '', $param, 1);
 
 		//$result1=$extrafields->addExtraField('chiffrage_myattr1', "New Attr 1 label", 'boolean', 1,  3, 'thirdparty',   0, 0, '', '', 1, '', 0, 0, '', '', 'chiffrage@chiffrage', '$conf->chiffrage->enabled');
 		//$result2=$extrafields->addExtraField('chiffrage_myattr2', "New Attr 2 label", 'varchar', 1, 10, 'project',      0, 0, '', '', 1, '', 0, 0, '', '', 'chiffrage@chiffrage', '$conf->chiffrage->enabled');